requirements.txt
asgiref==3.2.10
DateTime==4.3
Django==3.1.1
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.1
gunicorn==20.0.4
mysqlclient==2.0.1
pymssql==2.1.4
PyMySQL==0.10.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
zope.interface==5.1.0
ortools~=7.8.7959
